+++ cli/smartcontract/smart_contract.go
@@ -11,13 +11,13 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/CityOfZion/neo-go/pkg/compiler"
 	"github.com/CityOfZion/neo-go/pkg/crypto/hash"
 	"github.com/CityOfZion/neo-go/pkg/crypto/keys"
 	"github.com/CityOfZion/neo-go/pkg/rpc"
 	"github.com/CityOfZion/neo-go/pkg/smartcontract"
 	"github.com/CityOfZion/neo-go/pkg/util"
 	"github.com/CityOfZion/neo-go/pkg/vm"
-	"github.com/CityOfZion/neo-go/pkg/vm/compiler"
 	"github.com/go-yaml/yaml"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli"
