+++ pkg/compiler/analysis.go
