+++ pkg/compiler/emit.go
