+++ pkg/compiler/func_scope.go
